requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.20.0
numpy>=1.24.0
reportlab>=4.0.0
flask>=3.0.0
